refactor(aws): extract the shared KMS signature key store factory
The issuer and verifier signature key stores differ only in their alias
namespace, key tag and key-not-found error code, so move the KMS interaction
(generate, import, describe, sign) into createKmsSignatureKeyStore() and leave
kmsIssuerSignatureKeyStore as a thin wrapper over it. Behaviour is unchanged:
the existing issuer spec passes without modification.
